==> 02_pins.qmd <==
---
title: "Gecko datatset into pins"
format: 
  html:
    page-layout: full
    fontsize: smaller
editor: visual
execute:
  warning: false
---

# Pull data

```{r}
library(tidyverse)
library(pins)
library(finalfit)
source("01_pull.R")
#save(patient_data_orig, file = "patient_data_orig.rda")
#load("patient_data_orig.rda")
labels_keep = extract_variable_label(patient_data_orig)
source("02_join_world_bank.R")
source("03_cut_collapse.R")
```

# Make NAs explicit

```{r}
patient_data = patient_data %>% 
  mutate(across(where(is.factor), ~fct_na_value_to_level(., "Missing")))
```

# Cleaning
```{r}
source("04_cleaning.R")
```


# Select variables for Shiny app, reorder for dropdown menu

```{r}
patient_data = ff_relabel(patient_data, labels_keep)

appdata = patient_data %>%
  select(-record_id, -age_years, -admission_prior, -post30_los,
         -pre_symp_adm_day,
         -pre_diag_dec_day,
         -pre_dec_op_day,
         -pt_comorbid,
         -pre_img_finding,
         -pre_img_finding_cbd,
         -op_anaes,
         -iso2) %>% 
  relocate(wb, .after = period) %>% 
  relocate(age.groups, .after = wb) %>% 
  relocate(gender, .after = age.groups) %>%
  relocate(admission_prior.groups, .after = frailty) %>%
  relocate(starts_with("op_"), .after = hist_ac) %>% 
  relocate(all_of(c("pre_symp_adm_day.groups",
                    "pre_diag_dec_day.groups",
                    "pre_dec_op_day.groups")), .after = hist_ac)


 # allvars = appdata %>%
 #  finalfit::extract_labels() %>%
 #  select(vname, vfill) %$%
 #  setNames(as.list(vname), vfill)

updated_date = format(Sys.time(), format = "%d-%B %Y")
```

# Create aggregated dataset for shinylive

```{r}
gecko_aggre = patient_data %>% 
  drop_na() %>% 
  count(wb, gender, bmi, op_approach, pre_urgency, sort = TRUE) %>% 
  filter(n > 4)
```

# Write pin

```{r}

gecko_appdata = appdata

board = board_connect()
#board %>% pin_write(patient_data)
board %>% pin_write(gecko_appdata)
board %>% pin_write(gecko_aggre)

```

# Usage

```{r, eval = FALSE}
library(pins)
board = board_connect()
patient_data = pin_read(board, "rots/gecko_appdata")
patient_aggre = pin_read(board, "rots/gecko_aggre")
date_updated = pin_meta(board, "rots/gecko_appdata")$created
```
